+++ lua/kitty-scrollback/health.lua
@@ -15,7 +15,7 @@ end
 local function check_kitty_remote_control()
   vim.health.start('kitty-scrollback: Kitty remote control')
   local cmd = {
-    'kitty',
+    p.kitty_data.kitty_path,
     '@',
     'ls',
   }
@@ -206,8 +206,9 @@ local function check_kitty_debug_config()
   local kitty_debug_config_kitten =
     vim.api.nvim_get_runtime_file('python/kitty_debug_config.py', false)[1]
   local debug_config_log = vim.fn.stdpath('data') .. '/kitty-scrollback.nvim/debug_config.log'
-  local result =
-    vim.system({ 'kitty', '@', 'kitten', kitty_debug_config_kitten, debug_config_log }):wait()
+  local result = vim
+    .system({ p.kitty_data.kitty_path, '@', 'kitten', kitty_debug_config_kitten, debug_config_log })
+    :wait()
   if result.code == 0 then
     if vim.fn.filereadable(debug_config_log) then
       vim.health.ok(table.concat(vim.fn.readfile(debug_config_log), '\n   '))