+++ lua/kitty-scrollback/kitty_commands.lua
@@ -121,8 +121,12 @@ end
 
 M.get_text_term = function(kitty_data, get_text_opts, on_exit_cb)
   local esc = vim.fn.eval([["\e"]])
-  local kitty_get_text_cmd =
-    string.format([[kitty @ get-text --match="id:%s" %s]], kitty_data.window_id, get_text_opts)
+  local kitty_get_text_cmd = string.format(
+    [[%s @ get-text --match="id:%s" %s]],
+    p.kitty_data.kitty_path,
+    kitty_data.window_id,
+    get_text_opts
+  )
   local sed_cmd = string.format(
     [[sed -E ]]
       .. [[-e 's/%s\[\?25.%s\[.*;.*H%s\[.*//g' ]] -- remove control sequence added by --add-cursor flag
@@ -132,7 +136,7 @@ M.get_text_term = function(kitty_data, get_text_opts, on_exit_cb)
     esc,
     esc
   )
-  local flush_stdout_cmd = [[kitty +runpy 'sys.stdout.flush()']]
+  local flush_stdout_cmd = p.kitty_data.kitty_path .. [[ +runpy 'sys.stdout.flush()']]
   -- start to set title but do not complete see https://github.com/kovidgoyal/kitty/issues/719#issuecomment-952039731
   local start_set_title_cmd = string.format([[printf '%s]2;']], esc)
   local full_cmd = kitty_get_text_cmd
@@ -245,7 +249,7 @@ M.send_paste_buffer_text_to_kitty_and_quit = function(execute_command)
   end
 
   system_handle_error({
-    'kitty',
+    p.kitty_data.kitty_path,
     '@',
     'send-text',
     '--match=id:' .. p.kitty_data.window_id,
@@ -256,7 +260,7 @@ end
 
 M.list_kitty_windows = function()
   return system_handle_error({
-    'kitty',
+    p.kitty_data.kitty_path,
     '@',
     'ls',
   })
@@ -267,7 +271,7 @@ M.close_kitty_loading_window = function(ignore_error)
     local winid = p.kitty_loading_winid
     p.kitty_loading_winid = nil
     return system_handle_error({
-      'kitty',
+      p.kitty_data.kitty_path,
       '@',
       'close-window',
       '--match=id:' .. winid,
@@ -278,7 +282,7 @@ end
 
 M.signal_winchanged_to_kitty_child_process = function()
   system_handle_error({
-    'kitty',
+    p.kitty_data.kitty_path,
     '@',
     'signal-child',
     'SIGWINCH',
@@ -290,7 +294,7 @@ M.signal_term_to_kitty_child_process = function(force)
     vim.cmd.quitall({ bang = true })
   else
     system_handle_error({
-      'kitty',
+      p.kitty_data.kitty_path,
       '@',
       'signal-child',
       'SIGTERM',
@@ -303,7 +307,7 @@ M.open_kitty_loading_window = function(env)
     M.close_kitty_loading_window(true)
   end
   local kitty_cmd = vim.list_extend({
-    'kitty',
+    p.kitty_data.kitty_path,
     '@',
     'launch',
     '--type',
@@ -332,7 +336,7 @@ end
 M.get_kitty_colors = function(kitty_data, ignore_error, no_window_id)
   local match = no_window_id and nil or '--match=id:' .. kitty_data.window_id
   local ok, result = system_handle_error({
-    'kitty',
+    p.kitty_data.kitty_path,
     '@',
     'get-colors',
     match,
@@ -351,7 +355,7 @@ end
 
 M.send_text_to_clipboard = function(text)
   return system_handle_error({
-    'kitty',
+    p.kitty_data.kitty_path,
     '+kitten',
     'clipboard',
     '/dev/stdin',